added routes filder and bcrypt library
