+++ data/users.json
@@ -1,0 +1,12 @@
+[
+  {
+    "name": "joema76G76G76Gmwwwwwa",
+    "username": "kimmy123",
+    "password": "$2b$10$TphfljqOSJYM7ENMIL85NezXf5fyRuYCy7cCHaQZl9v9.GhWv9Xtu"
+  },
+  {
+    "name": "kimmy123",
+    "username": "kimmy123",
+    "password": "$2b$10$/XfgzHXelhqf/IssZznbyOZkqz3wca7rXLAY/7J5qaOIlYvYldID."
+  }
+]
